requirements.txt
# 智能数据库分析系统 - Python依赖包
# 修复版本 - 移除了内置模块 sqlite3

# AI 和数据处理核心包
anthropic>=0.7.0
pandas>=1.3.0

# 数据处理和可视化
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Excel支持
openpyxl>=3.0.0
xlrd>=2.0.0

# 配置管理
python-dotenv>=1.0.0

# 数据库和存储
# sqlite3 是Python内置模块，无需安装

# 界面和工具包
# numpy>=1.21.0
# openpyxl>=3.0.0
# xlrd>=2.0.0

# 可选：数据可视化支持
# matplotlib>=3.5.0
# seaborn>=0.11.0

# 可选：更好的JSON处理
ujson>=5.0.0

# 可选：HTTP客户端（用于健康检查）
requests>=2.28.0

# 开发和调试工具（可选）
# python-dateutil>=2.8.0
# pytz>=2022.1

# 新增依赖
sqlite3-api>=0.1.0
tkinter-utils>=0.1.0